+++ crates/sui-indexer-alt-graphql/src/api/types/event/mod.rs
@@ -9,9 +9,9 @@ use async_graphql::Object;
 use prost_types::FieldMask;
 use serde::Deserialize;
 use serde::Serialize;
-use sui_indexer_alt_reader::alpha_ledger_grpc_reader::AlphaLedgerGrpcReader;
-use sui_indexer_alt_reader::alpha_ledger_grpc_reader::StreamPage;
 use sui_indexer_alt_reader::kv_loader::KvLoader;
+use sui_indexer_alt_reader::ledger_grpc_reader::LedgerGrpcReader;
+use sui_indexer_alt_reader::ledger_grpc_reader::StreamPage;
 use sui_rpc::field::FieldMaskUtil;
 use sui_rpc::proto::sui::rpc::v2;
 use sui_rpc_cursor::CursorKind;
@@ -183,14 +183,14 @@ impl Event {
         filter: EventFilter,
     ) -> Result<StreamConnection<Event>, RpcError> {
         query_limits::rich::debit(ctx)?;
-        let reader: &AlphaLedgerGrpcReader = ctx.data()?;
+        let reader: &LedgerGrpcReader = ctx.data()?;
         Self::paginate_grpc(reader, scope, page, filter).await
     }
 
     /// Serve event pagination by streaming gRPC. Returns pages that may be partially filled,
     /// with valid cursors if there are more pages to paginate through.
     async fn paginate_grpc(
-        reader: &AlphaLedgerGrpcReader,
+        reader: &LedgerGrpcReader,
         scope: Scope,
         page: Page<CEvent>,
         filter: EventFilter,
@@ -430,7 +430,7 @@ mod tests {
     use fastcrypto::encoding::Base58;
     use fastcrypto::encoding::Base64 as B64;
     use fastcrypto::encoding::Encoding;
-    use sui_indexer_alt_reader::alpha_ledger_grpc_reader::PageItem;
+    use sui_indexer_alt_reader::ledger_grpc_reader::PageItem;
     use sui_types::base_types::ObjectID;
     use sui_types::parse_sui_struct_tag;
