+++ crates/sui-indexer-alt-graphql/src/api/types/transaction/mod.rs
@@ -11,11 +11,11 @@ use async_graphql::Object;
 use async_graphql::connection::Connection;
 use fastcrypto::encoding::Base58;
 use fastcrypto::encoding::Encoding;
-use sui_indexer_alt_reader::alpha_ledger_grpc_reader::AlphaLedgerGrpcReader;
-use sui_indexer_alt_reader::alpha_ledger_grpc_reader::StreamPage;
 use sui_indexer_alt_reader::kv_loader::KvLoader;
 use sui_indexer_alt_reader::kv_loader::TransactionContents as NativeTransactionContents;
 use sui_indexer_alt_reader::ledger_grpc_reader::CheckpointedTransaction;
+use sui_indexer_alt_reader::ledger_grpc_reader::LedgerGrpcReader;
+use sui_indexer_alt_reader::ledger_grpc_reader::StreamPage;
 use sui_rpc::proto::sui::rpc::v2;
 use sui_rpc_cursor::CursorKind;
 use sui_rpc_cursor::CursorToken;
@@ -323,7 +323,7 @@ impl Transaction {
         filter: TransactionFilter,
     ) -> Result<StreamConnection<Transaction>, RpcError> {
         query_limits::rich::debit(ctx)?;
-        let reader: &AlphaLedgerGrpcReader = ctx.data()?;
+        let reader: &LedgerGrpcReader = ctx.data()?;
         Self::paginate_grpc(reader, scope, page, filter).await
     }
 
@@ -334,7 +334,7 @@ impl Transaction {
     /// with fields hydrated lazily through the index) rather than through the `ctx`-driven
     /// [`Self::paginate`] wrapper.
     async fn paginate_grpc(
-        reader: &AlphaLedgerGrpcReader,
+        reader: &LedgerGrpcReader,
         scope: Scope,
         page: Page<CTransaction>,
         filter: TransactionFilter,
@@ -371,7 +371,7 @@ impl Transaction {
     /// scans forward to whatever is indexed. Items are returned in scan order (descending when
     /// paginating from the back).
     pub(crate) async fn scan_grpc(
-        reader: &AlphaLedgerGrpcReader,
+        reader: &LedgerGrpcReader,
         cp_bounds: impl RangeBounds<u64>,
         page: &Page<CTransaction>,
         filter: &TransactionFilter,
@@ -593,8 +593,8 @@ mod tests {
     use super::*;
     use crate::pagination::PageLimits;
     use async_graphql::connection::CursorType;
-    use sui_indexer_alt_reader::alpha_ledger_grpc_reader::PageItem;
     use sui_indexer_alt_reader::kv_loader::ExecutedTransactionData;
+    use sui_indexer_alt_reader::ledger_grpc_reader::PageItem;
     use sui_rpc_cursor::CursorKind;
     use sui_types::base_types::random_object_ref;
     use sui_types::effects::TransactionEffects as NativeTransactionEffects;
