+++ crates/sui-indexer-alt-reader/src/lib.rs
@@ -1,7 +1,6 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod alpha_ledger_grpc_reader;
 pub mod checkpoints;
 pub mod consistent_reader;
 pub mod cp_sequence_numbers;
